requirements.txt
# PPT复刻Skill - Python依赖清单
# 安装命令: pip install -r requirements.txt

# 核心依赖
python-pptx>=0.6.21,<1.1.0    # PPT解析与生成核心库
lxml>=4.9.0                   # XML解析（python-pptx的依赖）
requests>=2.28.0,<3.0.0       # HTTP请求（飞书API调用）

# 可选依赖（Webhook服务器模式）
flask>=2.0.0,<3.0.0           # Webhook服务器（可选）

# 开发依赖（测试和调试）
# pytest>=7.0.0               # 单元测试框架
# pytest-cov>=4.0.0           # 测试覆盖率
# black>=22.0.0               # 代码格式化
# flake8>=5.0.0               # 代码风格检查

# 注意：
# 1. python-pptx 1.0+ 版本引入了破坏性变更，请确保版本兼容
# 2. 如果使用Python 3.12+，请确保使用 python-pptx >= 0.6.23
# 3. requests 库用于飞书API调用，保持最新稳定版
